+++ release/ia64/mkisoimages.sh
@@ -4,7 +4,7 @@
 # Author: Jordan K Hubbard
 # Date:   22 June 2001
 #
-# $FreeBSD: release/7.0.0/release/ia64/mkisoimages.sh 141066 2005-01-30 21:10:52Z kensmith $
+# $FreeBSD$
 #
 # This script is used by release/Makefile to build the (optional) ISO images
 # for a FreeBSD release.  It is considered architecture dependent since each
@@ -41,56 +41,41 @@ LABEL=$1; shift
 NAME=$1; shift
 BASE=$1; shift
 
-MKISOFS=mkisofs
-MKISOFS_PKG=cdrtools
-MKISOFS_PORT=/usr/ports/sysutils/${MKISOFS_PKG}
-
-if ! which ${MKISOFS} > /dev/null; then
-    echo -n "${MKISOFS}(8) does not exist: "
-    if [ -f ${MKISOFS_PORT}/Makefile ]; then
-	echo building the port...
-	if ! (cd ${MKISOFS_PORT} && make install BATCH=yes && make clean); then
-	    echo "error: cannot build ${MKISOFS}(8). Bailing out..."
-	    exit 2
-	fi
-    else
-	echo fetching the package...
-	if ! pkg_add -r ${MKISOFS_PKG}; then
-	    echo "error: cannot fetch ${MKISOFS}(8). Bailing out..."
-	    exit 2
-	fi
-    fi
-fi
-
 EFIPART=efipart.sys
 
 # To create a bootable CD under EFI, the boot image should be an EFI
 # system partition.
 if [ $bootable = yes ]; then
-    EFISZ=32768
+    EFISZ=65536
     MNT=/mnt
-    dd if=/dev/zero of=$BASE/$EFIPART count=$EFISZ
-    md=`mdconfig -a -t vnode -f $BASE/$EFIPART`
+    dd if=/dev/zero of=$EFIPART count=$EFISZ
+    md=`mdconfig -a -t vnode -f $EFIPART`
     newfs_msdos -F 12 -S 512 -h 4 -o 0 -s $EFISZ -u 16 $md
     mount -t msdosfs /dev/$md $MNT
     mkdir -p $MNT/efi/boot $MNT/boot $MNT/boot/kernel
     cp -R $BASE/boot/defaults $MNT/boot
     cp $BASE/boot/kernel/kernel $MNT/boot/kernel
-    cp $BASE/boot/kernel/ispfw.ko $MNT/boot/kernel
+    if [ -s $BASE/boot/kernel/ispfw.ko ]; then
+	cp $BASE/boot/kernel/ispfw.ko $MNT/boot/kernel
+    fi
     cp $BASE/boot/device.hints $MNT/boot
     cp $BASE/boot/loader.* $MNT/boot
-    cp $BASE/boot/mfsroot.gz $MNT/boot
+    if [ -s $BASE/boot/mfsroot.gz ]; then
+	cp $BASE/boot/mfsroot.gz $MNT/boot
+    fi
     cp $BASE/boot/support.4th $MNT/boot
+    cp $BASE/boot/check-password.4th $MNT/boot
+    cp $BASE/boot/screen.4th $MNT/boot
     mv $MNT/boot/loader.efi $MNT/efi/boot/bootia64.efi
     umount $MNT
     mdconfig -d -u $md
-    BOOTOPTS="-b $EFIPART -no-emul-boot"
+    BOOTOPTS="-o bootimage=i386;$EFIPART -o no-emul-boot"
 else
     BOOTOPTS=""
 fi
 
-publisher="The FreeBSD Project.  http://www.freebsd.org/"
-
-$MKISOFS $BOOTOPTS -r -J -V $LABEL -publisher "$publisher" -o $NAME $BASE $*
-rm -f $BASE/$EFIPART
+echo "/dev/iso9660/$LABEL / cd9660 ro 0 0" > $BASE/etc/fstab
+makefs -t cd9660 $BOOTOPTS -o rockridge -o label=$LABEL $NAME $BASE $*
+rm $BASE/etc/fstab
+rm -f $EFIPART
 exit 0
